fix(pipeline): switch to BC Data Catalogue WFS for parcels and fix transit types (#11)
- Replace non-existent Richmond GeoHub ArcGIS URLs with ParcelMap BC WFS
- Fix ALR WFS bbox format (add EPSG:4326 suffix)
- Fix transit stop NUMERIC/float type mismatch with SQL casts
- Skip zoning/buildings ingestion (no public API available from Richmond BC)
- Successfully ingested: 9,626 parcels, 16 ALR boundaries, 1,020 transit stops

+++ backend/sync/ingest_transit.py
@@ -7,6 +7,7 @@ import io
 import zipfile
 
 import requests
+
 from backend.sync.config import DATABASE_URL, RICHMOND_BBOX, TRANSLINK_GTFS_URL
 from backend.sync.utils import get_db_connection
 
@@ -95,7 +96,8 @@ async def ingest_transit() -> int:
             await conn.execute(
                 """
                 INSERT INTO transit_stops (stop_name, route_type, lat, lng, geom)
-                VALUES ($1, $2, $3, $4, ST_SetSRID(ST_MakePoint($4, $3), 4326))
+                VALUES ($1, $2, $3::numeric, $4::numeric,
+                        ST_SetSRID(ST_MakePoint($4::float, $3::float), 4326))
                 """,
                 stop["stop_name"],
                 stop["route_type"],
